Fixed design of category list
