+++ package.json
@@ -54,7 +54,7 @@
     "@babel/preset-env": "^7.12.11",
     "@babel/preset-react": "^7.12.10",
     "@date-io/date-fns": "^1.3.13",
-    "@dmm-com/airone-apiclient-typescript-fetch": "^0.4.0",
+    "@dmm-com/airone-apiclient-typescript-fetch": "^0.5.0",
     "@dnd-kit/core": "^6.1.0",
     "@dnd-kit/modifiers": "^7.0.0",
     "@dnd-kit/sortable": "^8.0.0",
